update deleteUsuarioById method

+++ demo/src/main/java/com/buses/demo/service/UsuarioService.java
@@ -49,7 +49,9 @@ public class UsuarioService {
         return usuarioRepo.save(updated);
     }
 
-    public void deleteUsuarioById(long id){
+    public void deleteUsuarioById(long id) throws RecordNotFoundException{
+        if (usuarioRepo.existsById(id))
+            throw new RecordNotFoundException("Could not find usuario: " + id);
         usuarioRepo.deleteById(id);
     }
